pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qc-rule-paradox-scanner"
version = "0.1.0"
description = "自动扫描医院多套质控规则体系之间的逻辑冲突，生成冲突图谱和优先级消解建议"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Hospital Automation", email = "automation@example.com"}
]
keywords = ["hospital", "quality-control", "rule-conflict", "drg", "medical"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-docx>=1.1.0",
    "pdfplumber>=0.10.3",
    "jieba>=0.42.1",
    "openai>=1.12.0",
    "anthropic>=0.21.0",
    "jinja2>=3.1.3",
    "rich>=13.7.1",
    "typer>=0.12.3",
    "plotly>=5.19.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
qc-scan = "qc_rule_paradox_scanner.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
